Split UI related to file creation from filesel.[cpp|h] into filecreate.[cpp|h]
