+++ scripts/src/mame/frontend.lua
@@ -98,6 +98,8 @@ files {
 	MAME_DIR .. "src/frontend/mame/ui/filemngr.h",
 	MAME_DIR .. "src/frontend/mame/ui/filesel.cpp",
 	MAME_DIR .. "src/frontend/mame/ui/filesel.h",
+	MAME_DIR .. "src/frontend/mame/ui/filecreate.cpp",
+	MAME_DIR .. "src/frontend/mame/ui/filecreate.h",
 	MAME_DIR .. "src/frontend/mame/ui/floppycntrl.cpp",
 	MAME_DIR .. "src/frontend/mame/ui/floppycntrl.h",
 	MAME_DIR .. "src/frontend/mame/ui/imgcntrl.cpp",
